djangoProject/requirements.txt
django-cors-headers
djangorestframework
aniso8601==8.0.0
beautifulsoup4
certifi==2020.6.20
chardet==3.0.4
dlib
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
folium==0.12.1
googlemaps==4.4.5
gunicorn==20.0.4
html5lib
icecream==2.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
JPype1
jsonify==0.5
jupyter==1.0.0
konlpy==0.5.2
lxml
matplotlib
MarkupSafe==1.1.1
mccabe
mysql-connector
mysqlclient
nltk
numpy
pandas
pandas-datareader==0.9.0
pep8==1.7.1
plotly==5.0.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.24.0
seaborn
selenium==3.141.0
six==1.15.0
scikit-learn
SQLAlchemy==1.4.39
tensorflow==2.9.1
tensorflow-gpu==2.9.1
tensorflow-estimator==2.9.0
tensorflow_datasets
tqdm
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
torch
imbalanced-learn
tweepy==3.10.0
wordcloud
dataclasses==0.6
torchvision
opencv-python~=4.6.0.66
